refactor(reminders): split up AlarmManagerService
AlarmManagerService has acquired three responsibilities:

1. It is a BroadcastReceiver because it listens to snooze intents
2. It has the core functionality for scheduling review reminder notifications
3. It is supposed to be the centralized point for handling AlarmManager in AnkiDroid

This is too much responsibility for one file. This commit breaks it up into SnoozeService, ReviewReminderAlarmManager, and AlarmManagement respectively for the three responsibilities above.

Additionally, I found a lot of inconsistent logic scattered throughout the codebase for dealing with AlarmManager. I've centralized it all in AlarmManagement.

In the past, I found that AlarmManagerService, and specifically its logic relating to snoozing, were the hardest parts of the review reminders code to understand. I hope to improve on this through this change and an incoming documentation PR.
