+++ AnkiDroid/src/main/AndroidManifest.xml
@@ -638,7 +638,7 @@
             android:exported="false"
             />
         <receiver
-            android:name=".services.AlarmManagerService"
+            android:name=".services.SnoozeService"
             android:enabled="true"
             android:exported="false"
             />
