+++ AnkiDroid/src/main/java/com/ichi2/widget/DayRolloverAlarm.kt
@@ -24,7 +24,6 @@ import android.content.Intent.ACTION_TIMEZONE_CHANGED
 import android.content.Intent.ACTION_TIME_CHANGED
 import androidx.annotation.VisibleForTesting
 import androidx.core.app.PendingIntentCompat
-import androidx.core.content.getSystemService
 import anki.collection.opChanges
 import com.ichi2.anki.CollectionManager.withOpenColOrNull
 import com.ichi2.anki.common.android.AnkiBroadcastReceiver
@@ -36,7 +35,7 @@ import com.ichi2.anki.launchCatching
 import com.ichi2.anki.libanki.EpochMilliseconds
 import com.ichi2.anki.libanki.sched.Scheduler
 import com.ichi2.anki.observability.ChangeManager
-import com.ichi2.anki.services.AlarmManagerService
+import com.ichi2.utils.AlarmManagement
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import timber.log.Timber
@@ -87,15 +86,19 @@ class DayRolloverAlarm : AnkiBroadcastReceiver() {
         internal suspend fun scheduleNextInternal(context: Context) {
             val cutoffMs = nextFutureCutoffMs() ?: return
             val pendingIntent = context.buildPendingIntent(ACTION_ROLLOVER) ?: return
-            val alarmManager = context.getSystemService<AlarmManager>() ?: return
             // TODO: This uses setWindow with a 10 minute window, so likely fires at 04:10, rather than 04:00
             // Consider requesting SCHEDULE_EXACT_ALARM,
-            alarmManager.setWindowSafe(
-                type = AlarmManager.RTC_WAKEUP,
-                windowStartMillis = cutoffMs,
-                windowLengthMillis = AlarmManagerService.WINDOW_LENGTH_MS,
-                operation = pendingIntent,
-            )
+            AlarmManagement.useAlarmManager(context) { alarmManager ->
+                alarmManager.setWindow(
+                    AlarmManager.RTC_WAKEUP,
+                    cutoffMs,
+                    AlarmManagement.WINDOW_LENGTH_MS,
+                    pendingIntent,
+                )
+                val delta = (cutoffMs - TimeManager.time.intTimeMS()).milliseconds
+                Timber.i("scheduled day rollover")
+                Timber.d("rollover alarm scheduled for %s (in %s)", Date(cutoffMs), delta)
+            }
         }
 
         /**
@@ -121,26 +124,6 @@ class DayRolloverAlarm : AnkiBroadcastReceiver() {
                 false,
             )
     }
-}
-
-/**
- * [AlarmManager.setWindow] implementation which catches exceptions.
- */
-private fun AlarmManager.setWindowSafe(
-    type: Int,
-    windowStartMillis: EpochMilliseconds,
-    windowLengthMillis: Long,
-    operation: PendingIntent,
-) = try {
-    setWindow(type, windowStartMillis, windowLengthMillis, operation)
-    val delta = (windowStartMillis - TimeManager.time.intTimeMS()).milliseconds
-    Timber.i("scheduled day rollover")
-    Timber.d("rollover alarm scheduled for %s (in %s)", Date(windowStartMillis), delta)
-} catch (e: SecurityException) {
-    // #6332 - Too Many Alarms on Samsung Devices
-    Timber.w(e, "too many alarms — could not schedule alarm")
-} catch (e: Exception) {
-    Timber.w(e, "failed to schedule alarm")
 }
 
 /**
