+++ AnkiDroid/src/test/java/com/ichi2/anki/reviewreminders/ReviewReminderAlarmManagerTest.kt
@@ -1,10 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // SPDX-FileCopyrightText: Copyright (c) 2025 Eric Li <ericli3690@gmail.com>
 
-package com.ichi2.anki.services
+package com.ichi2.anki.reviewreminders
 
 import android.app.AlarmManager
-import android.app.NotificationManager
 import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
@@ -15,14 +14,10 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.RobolectricTest
 import com.ichi2.anki.common.time.MockTime
 import com.ichi2.anki.common.time.TimeManager
-import com.ichi2.anki.reviewreminders.ReviewReminder
-import com.ichi2.anki.reviewreminders.ReviewReminderId
-import com.ichi2.anki.reviewreminders.ReviewReminderScope
-import com.ichi2.anki.reviewreminders.ReviewReminderTime
-import com.ichi2.anki.reviewreminders.ReviewRemindersDatabase
 import com.ichi2.anki.services.NotificationService.Companion.EXTRA_REVIEW_REMINDER_ID
 import com.ichi2.anki.services.NotificationService.Companion.EXTRA_REVIEW_REMINDER_SCOPE
 import com.ichi2.anki.utils.ext.getParcelableCompat
+import com.ichi2.utils.AlarmManagement
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkStatic
@@ -36,11 +31,9 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.util.Calendar
-import kotlin.time.Duration
-import kotlin.time.Duration.Companion.minutes
 
 @RunWith(AndroidJUnit4::class)
-class AlarmManagerServiceTest : RobolectricTest() {
+class ReviewReminderAlarmManagerTest : RobolectricTest() {
     companion object {
         /**
          * Set the mock time to a consistent date and a specific time: noon in the local time zone (not UTC! this is required because
@@ -62,7 +55,6 @@ class AlarmManagerServiceTest : RobolectricTest() {
 
     private lateinit var context: Context
     private lateinit var alarmManager: AlarmManager
-    private lateinit var notificationManager: NotificationManager
     private lateinit var reviewReminder: ReviewReminder
 
     @Before
@@ -70,9 +62,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
         super.setUp()
         context = mockk(relaxed = true)
         alarmManager = mockk(relaxed = true)
-        notificationManager = mockk(relaxed = true)
         every { context.getSystemService<AlarmManager>() } returns alarmManager
-        every { context.getSystemService<NotificationManager>() } returns notificationManager
 
         reviewReminder = ReviewReminder.createReviewReminder(time = ReviewReminderTime(20, 0))
         reviewReminder.latestNotifTime = mockTime.intTimeMS() // Ensure the reminder is ready to have its future instance scheduled
@@ -95,12 +85,12 @@ class AlarmManagerServiceTest : RobolectricTest() {
         expectedSchedulingTime.apply {
             set(Calendar.HOUR_OF_DAY, 20)
         }
-        AlarmManagerService.scheduleReviewReminderNotification(context, reviewReminder, attemptImmediateNotification = true)
+        ReviewReminderAlarmManager.scheduleReviewReminderNotification(context, reviewReminder, attemptImmediateNotification = true)
         verify {
             alarmManager.setWindow(
                 AlarmManager.RTC_WAKEUP,
                 expectedSchedulingTime.timeInMillis,
-                AlarmManagerService.WINDOW_LENGTH_MS,
+                AlarmManagement.WINDOW_LENGTH_MS,
                 any(),
             )
         }
@@ -115,12 +105,12 @@ class AlarmManagerServiceTest : RobolectricTest() {
             set(Calendar.HOUR_OF_DAY, 3)
             add(Calendar.DAY_OF_YEAR, 1)
         }
-        AlarmManagerService.scheduleReviewReminderNotification(context, pastTimeReviewReminder, attemptImmediateNotification = false)
+        ReviewReminderAlarmManager.scheduleReviewReminderNotification(context, pastTimeReviewReminder, attemptImmediateNotification = false)
         verify {
             alarmManager.setWindow(
                 AlarmManager.RTC_WAKEUP,
                 expectedSchedulingTime.timeInMillis,
-                AlarmManagerService.WINDOW_LENGTH_MS,
+                AlarmManagement.WINDOW_LENGTH_MS,
                 any(),
             )
         }
@@ -135,12 +125,16 @@ class AlarmManagerServiceTest : RobolectricTest() {
             set(Calendar.HOUR_OF_DAY, 12)
             add(Calendar.DAY_OF_YEAR, 1)
         }
-        AlarmManagerService.scheduleReviewReminderNotification(context, currentTimeReviewReminder, attemptImmediateNotification = true)
+        ReviewReminderAlarmManager.scheduleReviewReminderNotification(
+            context,
+            currentTimeReviewReminder,
+            attemptImmediateNotification = true,
+        )
         verify {
             alarmManager.setWindow(
                 AlarmManager.RTC_WAKEUP,
                 expectedSchedulingTime.timeInMillis,
-                AlarmManagerService.WINDOW_LENGTH_MS,
+                AlarmManagement.WINDOW_LENGTH_MS,
                 any(),
             )
         }
@@ -148,7 +142,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
 
     @Test
     fun `scheduleReviewReminderNotifications attempts immediate notification when flag is true`() {
-        AlarmManagerService.scheduleReviewReminderNotification(context, reviewReminder, attemptImmediateNotification = true)
+        ReviewReminderAlarmManager.scheduleReviewReminderNotification(context, reviewReminder, attemptImmediateNotification = true)
 
         val slot = slot<Intent>()
         verify(exactly = 1) {
@@ -162,7 +156,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
 
     @Test
     fun `scheduleReviewReminderNotifications does not attempt immediate notification when flag is false`() {
-        AlarmManagerService.scheduleReviewReminderNotification(context, reviewReminder, attemptImmediateNotification = false)
+        ReviewReminderAlarmManager.scheduleReviewReminderNotification(context, reviewReminder, attemptImmediateNotification = false)
 
         verify(exactly = 0) {
             context.sendBroadcast(any())
@@ -183,12 +177,12 @@ class AlarmManagerServiceTest : RobolectricTest() {
             )
         } returns pendingIntent
 
-        AlarmManagerService.unscheduleReviewReminderNotifications(context, reviewReminder)
+        ReviewReminderAlarmManager.unscheduleReviewReminderNotifications(context, reviewReminder)
         verify { alarmManager.cancel(pendingIntent) }
     }
 
     @Test
-    fun `scheduleAllNotifications schedules and fires reminders for all enabled reminders in database`() =
+    fun `scheduleAllEnabledReviewReminderNotifications schedules and fires reminders for all enabled reminders in database`() =
         runTest {
             val did1 = addDeck("Deck1")
             val did2 = addDeck("Deck2")
@@ -212,7 +206,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
                 )
             reviewReminders.forEach { ReviewRemindersDatabase.insertReminder(it) }
 
-            AlarmManagerService.scheduleAllNotifications(context)
+            ReviewReminderAlarmManager.scheduleAllEnabledReviewReminderNotifications(context)
 
             reviewReminders.forEach { reminder ->
                 val expectedSchedulingTime = mockTime.calendar().clone() as Calendar
@@ -230,7 +224,7 @@ class AlarmManagerServiceTest : RobolectricTest() {
                     alarmManager.setWindow(
                         AlarmManager.RTC_WAKEUP,
                         expectedSchedulingTime.timeInMillis,
-                        AlarmManagerService.WINDOW_LENGTH_MS,
+                        AlarmManagement.WINDOW_LENGTH_MS,
                         any(),
                     )
                 }
@@ -248,59 +242,4 @@ class AlarmManagerServiceTest : RobolectricTest() {
                     }.toSet()
             assertThat(firedReminderIds, equalTo(expectedFiredReminderIds))
         }
-
-    @Test
-    fun `triggering schedules snoozed notification and cancels clicked notification`() =
-        runTest {
-            ReviewRemindersDatabase.insertReminder(reviewReminder)
-
-            attemptSnooze(reviewReminder, 5.minutes)
-
-            verifyNotifSnoozed(5.minutes)
-            verifyPastNotifCleared(reviewReminder)
-        }
-
-    @Test
-    fun `triggering only clears past notif if review reminder is not in database`() =
-        runTest {
-            attemptSnooze(reviewReminder, 5.minutes)
-
-            verifyNoNotifSnoozed()
-            verifyPastNotifCleared(reviewReminder)
-        }
-
-    private suspend fun attemptSnooze(
-        reviewReminder: ReviewReminder,
-        delay: Duration,
-    ) {
-        AlarmManagerService.handleSnoozeReviewReminder(
-            context,
-            reviewReminder.id,
-            reviewReminder.scope,
-            snoozeIntervalInMinutes = delay.inWholeMinutes.toInt(),
-        )
-    }
-
-    private fun verifyNotifSnoozed(delay: Duration) {
-        verify(exactly = 1) {
-            alarmManager.setWindow(
-                AlarmManager.RTC_WAKEUP,
-                mockTime.intTimeMS() + delay.inWholeMilliseconds,
-                AlarmManagerService.WINDOW_LENGTH_MS,
-                any(),
-            )
-        }
-    }
-
-    private fun verifyNoNotifSnoozed() {
-        verify(exactly = 0) {
-            alarmManager.setWindow(AlarmManager.RTC_WAKEUP, any(), any(), any())
-        }
-    }
-
-    private fun verifyPastNotifCleared(reviewReminder: ReviewReminder) {
-        verify(exactly = 1) {
-            notificationManager.cancel(NotificationService.REVIEW_REMINDER_NOTIFICATION_TAG, reviewReminder.id.value)
-        }
-    }
 }
